+++ tests/EntryPoint/EntryPointsBuilderTest.php
@@ -2,15 +2,18 @@
 
 namespace HeimrichHannot\EncoreBundle\Test\EntryPoint;
 
+use Contao\CoreBundle\Routing\ResponseContext\ResponseContext;
+use Contao\CoreBundle\Routing\ResponseContext\ResponseContextAccessor;
 use Contao\LayoutModel;
 use Contao\PageModel;
 use Contao\TestCase\ContaoTestCase;
-use HeimrichHannot\EncoreBundle\Asset\FrontendAsset;
 use HeimrichHannot\EncoreBundle\Collection\EntryCollection;
 use HeimrichHannot\EncoreBundle\EntryPoint\EntryPoint;
 use HeimrichHannot\EncoreBundle\EntryPoint\EntryPointBuilderFactory;
 use HeimrichHannot\EncoreBundle\EntryPoint\EntryPoints;
 use HeimrichHannot\EncoreBundle\EntryPoint\EntryPointsBuilder;
+use HeimrichHannot\EncoreBundle\Request\ResponseContext\Entry;
+use HeimrichHannot\EncoreBundle\Request\ResponseContext\EntryBag;
 use HeimrichHannot\TestUtilitiesBundle\Mock\ModelMockTrait;
 use HeimrichHannot\UtilsBundle\Util\ModelUtil;
 use HeimrichHannot\UtilsBundle\Util\Utils;
@@ -118,13 +121,17 @@ class EntryPointsBuilderTest extends ContaoTestCase
             ->method('model')
             ->willReturn($modelUtil);
 
-        $frontendAsset = new FrontendAsset();
-        $frontendAsset->addActiveEntrypoint('frontend-entry');
-        $frontendAsset->addActiveEntrypoint('missing-frontend-entry');
+        $responseContext = new ResponseContext();
+        $responseContext->add(
+            (new EntryBag())
+                ->addEntry(new Entry('frontend-entry', 'frontend', 'App'))
+                ->addEntry(new Entry('missing-frontend-entry', 'frontend', 'App'))
+        );
 
         $builder = new EntryPointsBuilder($utils, $entryCollection);
         $result = $builder
-            ->setFrontendAsset($frontendAsset)
+            ->setResponseContext($responseContext)
+            ->setCustomBag(null)
             ->setLayout($layout, 'layoutEntries')
             ->setPage($page, 'customEntries')
             ->build();
@@ -143,7 +150,7 @@ class EntryPointsBuilderTest extends ContaoTestCase
             array_keys($active)
         );
 
-        $this->assertSame(FrontendAsset::class, $all['frontend-entry']->origin);
+        $this->assertSame('frontend', $all['frontend-entry']->origin);
         $this->assertFalse($all['frontend-entry']->requiresCss);
         $this->assertTrue($all['layout-entry']->head);
         $this->assertTrue($all['layout-entry']->requiresCss);